DOM and DOM Manipulation
DOM(Document Object Model): Structured representation of HTML documents. Allows JS to access HTML elements and styles to manipulate them.

DOM is not a part of  JS
but how does it work in JS? DOM and DOM Method is part of web APIs. we can access from our JS code

document.querySelector('.selector')  : method that is available on the document element (use the selector exactly same selector which used in CSS)