Handling Keypress Event
Keyboard event is so-called global event because they do not happen on specific element. Global events usually listen on the whole document.

document.addEventListener() -> listening for events everywhere on document.

keydown : as soon as press down the key
keypress : continuously as we keep finger on the key
keyup : when lifting finger off the key